Rename dimensions package modules with underscores.
This serves two purposes:

- it avoids the module names being publicly exported as symbols by
  higher-level __init__.py files (this is already the practice used in
  'registry' and other newer daf_butler subpackages);

- because elements.py is _not_ being renamed yet, it makes the future
  commit on which it is not just renamed but rewritten easier to
  understand, because it'll look like a totally new file.

This commit is _just_ moving files and changing imports accordingly.
