+++ python/lsst/daf/butler/core/dimensions/_config.py
