+++ python/lsst/daf/butler/core/dimensions/_records.py
@@ -36,8 +36,8 @@ from ..timespan import Timespan, DatabaseTimespanRepresentation
 
 if TYPE_CHECKING:  # Imports needed only for type annotations; may be circular.
     from .elements import DimensionElement
-    from .coordinate import DataCoordinate
-    from .schema import DimensionElementFields
+    from ._coordinate import DataCoordinate
+    from ._schema import DimensionElementFields
 
 
 def _reconstructDimensionRecord(definition: DimensionElement, mapping: Dict[str, Any]) -> DimensionRecord:
@@ -54,7 +54,7 @@ def _subclassDimensionRecord(definition: DimensionElement) -> Type[DimensionReco
 
     For internal use by `DimensionRecord`.
     """
-    from .schema import DimensionElementFields, REGION_FIELD_SPEC
+    from ._schema import DimensionElementFields, REGION_FIELD_SPEC
     fields = DimensionElementFields(definition)
     slots = list(fields.standard.names)
     if definition.spatial:
@@ -136,7 +136,7 @@ class DimensionRecord:
                     kwargs.get("datetime_end"),
                 )
 
-        from .coordinate import DataCoordinate
+        from ._coordinate import DataCoordinate
         object.__setattr__(
             self,
             "dataId",
