+++ python/lsst/daf/butler/core/dimensions/_schema.py
